让索引等值查询简化，少一次索引查询

+++ src/record/rm_file_handle.cpp
@@ -84,7 +84,7 @@ void RmFileHandle::set_meta_ts(Transaction *txn, size_t log_idx, const Rid &rid,
 }
 
 // 热点行的更新操作只需保证串行安全即可 所以commit时就落地更新
-void RmFileHandle::set_hot_record(const Rid &rid, ColMeta &col_meta, int32_t delta) {
+void RmFileHandle::set_hot_record(const Rid &rid, ColMeta &col_meta, float delta) {
   RmPageHandle page_hdl = fetch_page_handle(rid.page_no);
 
   //   找行锁 加锁写数据
@@ -92,8 +92,8 @@ void RmFileHandle::set_hot_record(const Rid &rid, ColMeta &col_meta, int32_t del
     auto rec_latch_ptr = get_rec_latch(rid);
     std::unique_lock<std::shared_mutex> rec_lock(*rec_latch_ptr);
     char *old_val = page_hdl.get_slot_record(rid.slot_no) + col_meta.offset;
-    // 必为整型
-    int new_val = *(int *)old_val + delta;
+    // 必为浮点数
+    float new_val = *(float *)old_val + delta;
     memcpy(old_val, &new_val, col_meta.len);
   }
